メインモジュールのリファクタリング #19
コマンドラインオプションの定義関係のモジュール分離
