+++ tex/list.tex
@@ -16,6 +16,10 @@
 \inputminted[linenos, breaklines]{rust}{src/main.rs}
 \clearpage
 
+\section{コマンドラインオプションの定義 cmdline\_opt.rs}
+\inputminted[linenos, breaklines]{rust}{src/cmdline_opt.rs}
+\clearpage
+
 \section{ファイルシステムモジュール ssh\_filesystem.rs}
 \inputminted[linenos, breaklines]{rust}{src/ssh_filesystem.rs}
